Декомпозиция страницы поиска, создание роута для страницы поиска курсов и тренеров. обновление ссылок на страницы поиска в компонентах
